dashboardAdmin: added type definitions, detailed comments added, reorganised files, revised array map to avoid potential error -> when the constant is not an array
